worker/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.5
pydantic-settings==2.7.1
python-dotenv==1.0.1
SQLAlchemy==2.0.37
PyMySQL==1.1.1
requests==2.32.3
msal==1.31.1
cryptography==44.0.0
